+++ src/runtime/components/CldSource.vue
@@ -1,0 +1,15 @@
+
+  <script>
+  export default {
+    props: {
+      srcset: String,
+      media: String,
+    },
+  };
+  </script>
+
+<template>
+    <source :srcset="srcset" :media="media" />
+  </template>
+  
+
